Hide the hover-appearing close button
My friends to whom I've recommended Shout are all accidentally leaving
channels. This is because in the default theme, the channel close button
will only show up when the mouse pointer is on the channel name, and
beginners don't know that they exist at all. Even after they learn the
existence of such button, it takes more unwanted channel leavings to
finally be trained to avoid it.

This commit disables the hover-appearing close button by default. Users
who want the button can easily bring it back by adding to their theme:

    #sidebar .close { display: block; }

It is intended as a simpler alternative to adding a prompt message
dialogue, previously discussed in the pull request [558].

[558]: https://github.com/erming/shout/pull/558